From gerrit:
Author: Benjamin Kaduk <kaduk@mit.edu>
AuthorDate: 2014-04-07 17:55:09 -0400
Commit: Benjamin Kaduk <kaduk@mit.edu>
CommitDate: 2014-04-07 18:19:52 -0400

vol: Fix build with separate objdir

The volscan-main and volinfo-main source files are in the source
tree, not the object tree; refer to the objects in the Makefile
as dependencies, so that they will be picked up properly. The
objects will be made just fine by the implicit .c.o rule.

Change-Id: Ieec4b32cfbe5d260e1560a08d4ed8162720f9222