salvager patches for AFS/OSD
with this change the salvager and dasalvager become AFS/OSD aware.
The LWP-salvager will deny to salvage OSD-volumes while the pthreaded
dasalvager will try to load libdafsosd.so.1.xx. If this is not possible
it wil also deny to salvage the OSD-volume, otherwise it will salvage it
by means of the hooks in vol-salvage.c into the shared library.

Even if this is not yet a start to AFS/OSD this patch has the advantage
that a fileserver instance which was erranously started with the wrong
binaries on an OSD-fileserver will not destroy OSD-volumes anymore.