Let the cache manager support AFS/OSD
The AFS/OSD stuff is handled in afs_rcosd.c. However, to do this in a
uniform way parts of afs_bypasscache.c had to be integrated into
afs_fetchstore.c. Also afs_conn.c had to be extended to allow to
specify the service-id for a new connection.

By default the AFS/OSD support is switched off. Only if afsd is started
wilth "-libafsosd" it get's activated.