Integration of osddb in vlserver, ubik-extensions
The integration the OSDDB into the vlserver as a new service required
to change ubik that more than a single database can be handled.
The upgrade now allows up to four databases in a ubik database
instance.

This ubik extensions are not yet running in production with more
than a single database anywhere. Production cells sauch as RZG and
PSI still use the standalone osddbserver which listens on port 7012.
Therefore the vlserver started with "-libafsosd" still listens
additionally on port 7012 because all commands, file- and volservers
and rxosd still use this port for compatibility. This can be changed
once it has turned out that the ubik-extensions are stable.